refactor: js 파일들 모듈화시켜서 나누기
