+++ src/main/resources/public/js/store.js
@@ -1,0 +1,1 @@
+export let store = {}
